get rid of import cycle
