+++ internal/stream_fetch_helper.go
@@ -36,8 +36,8 @@ func GetLogsDstSettings(operationLogsDstEnvVariable string) (dstFolder string, e
 }
 
 // TODO : unit tests
-// downloadAndDecompressStream downloads, decompresses and writes stream to stdout
-func downloadAndDecompressStream(backup *Backup, writeCloser io.WriteCloser) error {
+// DownloadAndDecompressStream downloads, decompresses and writes stream to stdout
+func DownloadAndDecompressStream(backup *Backup, writeCloser io.WriteCloser) error {
 	for _, decompressor := range compression.Decompressors {
 		archiveReader, exists, err := TryDownloadFile(backup.BaseBackupFolder, getStreamName(backup.Name, decompressor.FileExtension()))
 		if err != nil {
